Match any user pkgs for headless sys user mode
In headless system user mode a package might not be installed in the
system user. Since the config list is only loaded and matched against
packages for the system user, we need to handle the case where a package
might exist on the system but not be installed for the system user.

Bug: 241469882
Test: atest NotificationListenersTest
Change-Id: I3e225047846658479d00d4089b8864860ef0c274
Merged-In: I3e225047846658479d00d4089b8864860ef0c274
(cherry picked from commit 19eb2d59a6ee6f1ed5d603bbed438b3febbf8145)