Fix some test-specific bubbles issues
First, the tapl test infra seems to launch bubble activities
in a different way from normal chat-app usage. The result is
that the activity doesn't become visible immediately (instead
waiting for onAttach). Because the activity-start is initiated
by WCT, it is expected to be ready after evaluating the WCT.
To fix this, just make sure that the shortcut-launch records
the EFFECTS_LIFECYCLE property so that WOC recalculates
the visibility during the transaction.

Secondly, the test infra seems to kill/restart sysui/shell
on every test, so it was regularly hitting binderdeath
handler. That handler was cleaning up playing transitions but
not the collecting one which caused the new shell instance
to receive onTransitionReady for a transition started with
the old shell instance. To fix this, just clean-up the
collectingTransition if one exists.

Bug: 253271977
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.bubble
Change-Id: I73f6bdf8c126cf212d9287222898fbe060c4022d