Stop using notification for IME switcher
InputMethodManagerService has had logic to show notifications when the
IME switcher icon cannot be shown on NavigationBar[1][2][3][4][5].

With this CL, such a fallback behavior is completely removed.  Devices
and form-factors that need a special handling need to come up with
their own solution rather than relying on this historical notification
mechanism.

 [1]: I2b9e955003c1f6703978824c947d5f322a039714
      7cfc0ed21a3fafabafb40b38a8589808ad1517af
 [2]: Id36c8c34159bea8b72557b40bcf024d401f580b6
      739b7b1eb2040e5a1b3da4a998702dd58ea337d0
 [3]: I34a95732ef3e7c20d6549b57230c11f0c3db04d6
      8e53a7fc9d48759ab999de3a3c80a37514af954b
 [4]: I58421143bee25be459f4f744688b37614a750051
      214dc68f7fcc3889a246a3a4cfa4ffa33586d6a4
 [5]: I6ad818fc9ef08991e24c5020498243f1fc417998
      9c98cd9ace10543498a2e2f0500b6558b471f5d6

Bug: 245989733
Test: atest CtsPermission2TestCases:ProtectedBroadcastsTest
Test: presubmit
Change-Id: I3b09d2b05ba9bde686b360b9f720b5d4258551c3