Simplify onProviderStateChangedOnHandler
Mostly a non-functional refactor, except for the addition
of a validity check when adding routes from a new provider.
Before this change the validity check was only made for
routes added by an existing provider, but not for routes
coming from a new provider. Under the assumption that this
was an oversight, we make the validity check for all routes.

Test: atest mediaroutertest CtsMediaBetterTogetherTestCases
Bug: 205124386
Change-Id: I4b4ca48756ac027e3c54b7c1000dfab05248682a