Clarify required permission in IMM#getInputMethodListAsUser()
This is a follow up CL to our previous CL [1], which exposed

  InputMethodManager#getInputMethodListAsUser(int userId) [2]

as an @TestApi.

One thing to clarify is that INTERACT_ACROSS_USERS_FULL permission is
required when and only when "userId" is different from the one of the
calling process.  To make it clear, this CL adds "conditional = true"
like other @RequiresPermission usages in InputMethodManager.

This is just about (Test) API annotation.  There must be no observable
behavior change.

 [1]: I4b73d9374d7a2ce4c9dfc48682a43d5d48b39042
      fb58ee92a033a9b9d33998379cf7d7489223fa0c
 [2]: I5b50b5fe143c74c87b331bda3e5bcc4d6248436e
      3c79a517a3113c2e89d9b7b8e64b8dde4c3e2222

Bug: 256239860
Test: presubmit
Change-Id: I14867e528d4292f1f8713b1d9b491fc5ed6c1abb