Improve module and apex printout
BinaryTransparencyService:
Make use of APEX digest as a binary measurement, which measures
only the changes to the content of the package, excluding the signing
block.

Additionally, this information is surfaced to the output of existing
`get apex_info` and `get module_info` commands.

The format of the output is also modified from previous iteration
to make the overall style and presentation more consistent for the
user.

PackageUtils:
Currently cryptographic digests that are hex encoded and encoded to
string are encoded in uppercase - this is not consistent with how
measurements for other binaries are stored and processed.

Thus, this change defaults the encoding for SHA-256 digest for
large files to be hex encoded and string encoded to lowercase.

Bug: 245941527

Test: manual.

Change-Id: I5b332214b516a71bded8dcc50883430608cde357