Remove an unused param from IMMS#showInputMethodPickerFromClient()
This is the 2nd attempt to submit this cleanup [1], after a revert [2] due to some build failure.

This is a follow up CL to our previous CL [3], which introduced

  InputMethodManagerService#showInputMethodPickerFromClient(
          @NonNull IInputMethodClient client,
          int auxiliarySubtypeMode, int displayId);

Somehow the "client" has never been used, which is fine because the
actual caller verification is done by requiring WRITE_SECURE_SETTINGS
permission.

Let's remove such an unused parameter for simplicity.

There must be no observable behavior change.

 [1]: I2f27160cbe7d83201c8f52dbb6dcb774a5b8ae43
      42d44afb5e3338350971c8ea3464855c08f6081c
 [2]: I343828ad83d50436211004753f1b9ba1b13b60ca
      e8772e8d2c11cf08da4a4d225feaa93a581d0da9
 [3]: Ic7d7c5a7ad8005a3fbd9d1c1b73e3c5a39a07001
      5bc5e2c7bc6bc4d3102acd92992268ccf3e2b5e9

Bug: 234882948
Bug: 256952514
Test: presubmit
Test: lunch aosp_bramble-userdebug && make -j
Change-Id: Ia536d8793f0022179dfafd1cbf21091537375992