On OTA, MainUser is given to an Admin user
Followup to Ib36f3b372f9bf33cbb097c4af63eb43515ac835b. For consistency, we give MainUser designation specifically to an Admin user. That way, for now, all MainUsers are indeed Admin users. We may not make this a permanent assumption going forward, but for now, it'd keep things consistent. And, after all, the "Main" user of the device probably is an Admin for the devices using HSUM so far.

Bug: 256624031
Change-Id: I46a26d3a8a93dc52cfe48963215a03aa8636f90c
Test: atest UserManagerTest