Use Dsp session and Software session in HotwordDetectionConnection
We use the HotwordDetectorSession to handle the Dsp and
Software trusted hotword detectors, it would be better to
separate them.

We create the DspTrustedHotwordDetectorSession to handle Dsp
trusted hotword detector, SoftwareTrustedHotwordDetectorSession
to handle software trusted hotword detector.

Currently they are just the simple classes, we will have another
change to move the methods of HotwordDetectorSession into
DspTrustedHotwordDetectorSession and
SoftwareTrustedHotwordDetectorSession.

Test: atest CtsVoiceInteractionTestCases
Bug: 241041976
Change-Id: I0d569eae319acd8260808aac545babd323f1657e